Route build assets to dist and reorganize scripts
