+++ 404.html
@@ -4,7 +4,7 @@
     <meta name="description" content="Página no encontrada - El Rincón de Ébano">
     <link rel="canonical" href="https://elrincondeebano.com/">
     <link rel="manifest" href="/app.webmanifest">
-    <script src="assets/js/csp.js"></script>
+    <script src="dist/js/csp.js"></script>
     <title>Página No Encontrada - El Rincón de Ébano</title>
     <link rel="stylesheet" href="assets/css/404.css">
     <link rel="preconnect" href="https://fonts.googleapis.com" crossorigin="">
@@ -17,7 +17,7 @@
 
     <!-- Preload critical assets -->
     <link rel="preload" href="assets/images/web/404.webp" as="image" type="image/webp">
-    <link rel="preload" href="assets/js/script.min.js" as="script">
+    <link rel="preload" href="dist/js/script.min.js" as="script">
 
 </head>
 <body>
