--- assets/js/gtag-init.js
@@ -1,7 +1,0 @@
-// Google Analytics initialization extracted from inline script
-window.dataLayer = window.dataLayer || [];
-function gtag(){ dataLayer.push(arguments); }
-gtag('js', new Date());
-// Prevent duplicate page views (matches previous config on index.html)
-gtag('config', 'G-H0YG3RTJVM', { 'send_page_view': false });
-
