+++ index.html
@@ -8,7 +8,7 @@
     <link rel="manifest" href="/app.webmanifest">
 
     <!-- Security -->
-    <script src="assets/js/csp.js?v=2" async=""></script>
+    <script src="dist/js/csp.js?v=2" async=""></script>
 
     <!-- Title and Favicon -->
     <title>El Rincón de Ébano</title>
@@ -30,30 +30,30 @@
     <link rel="dns-prefetch" href="//cdnjs.cloudflare.com">
     
     <!-- Preload critical assets -->
-    <link rel="preload" href="/assets/css/critical.min.css" as="style" fetchpriority="high">
+    <link rel="preload" href="/dist/css/critical.min.css" as="style" fetchpriority="high">
     <link rel="preload" href="/assets/images/web/logo.webp" as="image" type="image/webp" fetchpriority="high">
 <!-- Critical CSS -->
-    <link rel="stylesheet" href="assets/css/critical.min.css">
+    <link rel="stylesheet" href="dist/css/critical.min.css">
 
     <!-- Security - XSS Protection -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/dompurify/2.3.8/purify.min.js" integrity="sha384-AsiVBlzbaNOq8OOKcXm2ZVjjKJwiQ9UmzLwfetDjC74OMQdkb6vBHH5QRJH3x1SE" crossorigin="anonymous" defer=""></script>
 
     <!-- Preload Critical Assets -->
     <link rel="preload" href="assets/images/web/logo.webp" as="image" type="image/webp" fetchpriority="high">
-      <link rel="preload" href="assets/js/script.min.js" as="script">
+      <link rel="preload" href="dist/js/script.min.js" as="script">
 
     <!-- Deferred Non-Critical CSS -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" media="print" data-defer="" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&amp;family=Playfair+Display:wght@400;700&amp;display=swap" media="print" data-defer="">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css" media="print" data-defer="">
-    <link rel="stylesheet" href="assets/css/style.min.css?v=100" media="print" data-defer="">
+    <link rel="stylesheet" href="dist/css/style.min.css?v=100" media="print" data-defer="">
 
     <!-- Fallback for No JavaScript -->
     <noscript>
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
         <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&amp;family=Playfair+Display:wght@400;700&amp;display=swap">
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css">
-        <link rel="stylesheet" href="assets/css/style.min.css?v=100">
+        <link rel="stylesheet" href="dist/css/style.min.css?v=100">
     </noscript>
 
     <!-- Open Graph Tags -->
@@ -68,7 +68,7 @@
 
     <!-- Google tag (gtag.js) -->
     <script async="" src="https://www.googletagmanager.com/gtag/js?id=G-H0YG3RTJVM"></script>
-    <script src="assets/js/gtag-init.js" defer=""></script>
+    <script src="dist/js/gtag-init.js" defer=""></script>
 
     <header id="navbar-container" role="banner"></header>
 
@@ -124,9 +124,9 @@
     
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js" defer="" integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js" defer="" integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy" crossorigin="anonymous"></script>
-    <script src="assets/js/script.min.js" defer=""></script>
+    <script src="dist/js/script.min.js" defer=""></script>
 
-    <!-- Service Worker registration handled in assets/js/script.min.js -->
+    <!-- Service Worker registration handled in dist/js/script.min.js -->
