+++ package.json
@@ -7,8 +7,8 @@
     "test": "test"
   },
   "scripts": {
-    "build": "node build.js && node build-pages.js && node tools/inject-structured-data.js && node tools/generate-sitemap.js && node tools/inject-resource-hints.js",
-    "icons": "node generate-icons.js",
+    "build": "node tools/build.js && node tools/build-pages.js && node tools/inject-structured-data.js && node tools/generate-sitemap.js && node tools/inject-resource-hints.js",
+    "icons": "node tools/generate-icons.js",
     "images:variants": "node tools/generate-image-variants.js",
     "images:generate": "node tools/generate-images.mjs",
     "images:rewrite": "node tools/rewrite-images.mjs",
