+++ pages/energeticaseisotonicas.html
@@ -6,7 +6,7 @@
     <meta name="description" content="Explora nuestra selección de bebidas energéticas y isotónicas en El Rincón de Ébano.">
     <link rel="canonical" href="https://elrincondeebano.com/pages/energeticaseisotonicas.html">
     <link rel="manifest" href="/app.webmanifest">
-    <script src="../assets/js/csp.js"></script>
+    <script src="../dist/js/csp.js"></script>
     <title>Energéticas e Isotónicas - El Rincón de Ébano</title>
     <link rel="icon" type="image/x-icon" href="../assets/images/web/favicon.ico">
     <link rel="preconnect" href="https://cdn.jsdelivr.net">
@@ -21,22 +21,22 @@
     <link rel="dns-prefetch" href="//cdnjs.cloudflare.com">
     
     <!-- Preload critical assets -->
-    <link rel="preload" href="/assets/css/critical.min.css" as="style" fetchpriority="high">
+    <link rel="preload" href="/dist/css/critical.min.css" as="style" fetchpriority="high">
     <link rel="preload" href="/assets/images/web/logo.webp" as="image" type="image/webp" fetchpriority="high">
 <link rel="preload" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&family=Playfair+Display:wght@400;700&display=swap" as="style">
-    <link rel="stylesheet" href="../assets/css/critical.min.css">
+    <link rel="stylesheet" href="../dist/css/critical.min.css">
     <script src="https://cdnjs.cloudflare.com/ajax/libs/dompurify/2.3.8/purify.min.js" integrity="sha384-AsiVBlzbaNOq8OOKcXm2ZVjjKJwiQ9UmzLwfetDjC74OMQdkb6vBHH5QRJH3x1SE" crossorigin="anonymous"></script>
     <link rel="preload" href="../assets/images/web/logo.webp" as="image" type="image/webp">
-    <link rel="preload" href="../assets/js/script.min.js" as="script">
+    <link rel="preload" href="../dist/js/script.min.js" as="script">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" media="print" data-defer>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&family=Playfair+Display:wght@400;700&display=swap" media="print" data-defer>
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css" media="print" data-defer>
-    <link rel="stylesheet" href="../assets/css/style.min.css" media="print" data-defer>
+    <link rel="stylesheet" href="../dist/css/style.min.css" media="print" data-defer>
     <noscript>
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
         <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&family=Playfair+Display:wght@400;700&display=swap">
         <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css">
-        <link rel="stylesheet" href="../assets/css/style.min.css">
+        <link rel="stylesheet" href="../dist/css/style.min.css">
     </noscript>
     <meta property="og:title" content="El Rincón de Ébano - Energéticas e Isotónicas">
     <meta property="og:description" content="Explora nuestra selección de bebidas energéticas y isotónicas en El Rincón de Ébano.">
@@ -89,6 +89,6 @@
 
     <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js" integrity="sha384-oBqDVmMz9ATKxIep9tiCxS/Z9fNfEXiDAYTujMAeBAsjFuCZSmKbSSUnQlmh/jp3" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js" integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy" crossorigin="anonymous"></script>
-    <script src="../assets/js/script.min.js" defer></script>
+    <script src="../dist/js/script.min.js" defer></script>
 </body>
 </html>
