+++ service-worker.js
@@ -15,9 +15,9 @@ const CACHE_CONFIG = {
         '/',
         '/index.html',
         '/404.html',
-        '/assets/css/style.min.css',
-        '/assets/css/critical.min.css',
-        '/assets/js/script.min.js',
+        '/dist/css/style.min.css',
+        '/dist/css/critical.min.css',
+        '/dist/js/script.min.js',
         '/assets/images/web/logo.webp',
         '/assets/images/web/icon-192.png',
         '/assets/images/web/icon-512.png',
